refactor: moved search_parameters from networking folder to report folder
