+++ src/gui/types/message.rs
@@ -5,8 +5,8 @@ use crate::gui::pages::types::running_page::RunningPage;
 use crate::gui::pages::types::settings_page::SettingsPage;
 use crate::gui::styles::types::gradient_type::GradientType;
 use crate::networking::types::host::Host;
-use crate::networking::types::search_parameters::SearchParameters;
 use crate::notifications::types::notifications::Notification;
+use crate::report::types::search_parameters::SearchParameters;
 use crate::utils::types::file_info::FileInfo;
 use crate::utils::types::web_page::WebPage;
 use crate::{ChartType, IpVersion, Language, Protocol, ReportSortType, StyleType};
