+++ src/networking/types/mod.rs
@@ -15,7 +15,6 @@ pub mod my_link_type;
 pub mod packet_filters_fields;
 pub mod port_collection;
 pub mod protocol;
-pub mod search_parameters;
 pub mod service;
 pub mod service_query;
 pub mod traffic_direction;
