+++ src/report/types/mod.rs
@@ -1,3 +1,4 @@
 pub mod report_col;
 pub mod report_sort_type;
+pub mod search_parameters;
 pub mod sort_type;
